seprate constant strings in final class
